Google Cloud/Technical Expert Badge VertexAI/day2-3-Develop Sols using Model Garden models and APIs Challenge Lab/model-garden-lab-app/requirements.txt
streamlit
streamlit-extras
google-cloud-aiplatform
google-cloud-logging
openai
anthropic[vertex]